\name{HOPLSA}
\alias{PLSA}
\alias{PLSA_3way}
\alias{PLSA_4way}

\docType{package}
\title{\packageTitle{HOPLSA}}

\description{
\packageDescription{HOPLSA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{HOPLSA}
\packageIndices{HOPLSA}
See help for individual functions
}

\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{}

\examples{

}
